pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cd3-binder-design"
version = "0.1.0"
description = "Computational pipeline for designing CD3 binders for bispecific antibody therapeutics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Alec Nielsen"}
]
keywords = ["antibody", "bispecific", "CD3", "protein design", "drug discovery"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "biopython>=1.81",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "anarci>=1.3",
    "abnumber>=0.3.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "ablang2>=0.1.0",
]
modal = [
    "modal>=0.50.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "py3Dmol>=2.0.0",
]
all = [
    "cd3-binder-design[ml,modal,dev,notebooks]",
]

[project.urls]
Homepage = "https://github.com/alecnielsen/cd3-binder-design"
Repository = "https://github.com/alecnielsen/cd3-binder-design"

[project.scripts]
cd3-design = "src.pipeline.design_pipeline:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "modal*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
